requirements.txt
bitbucket-openapi
pyyaml
